Refactor results queries to get total questions from questions table, mane name optional and change route to user templateVars'
