+++ db/seeds/05_results.sql
@@ -1,21 +1,21 @@
-INSERT INTO results (quiz_id, user_id, total_questions, correct_answers, url_key) VALUES
-(1, 1, 4, 3, 'aB7xY9z2w1'),
-(2, 2, 4, 2, 'cDeFgH3i4j'),
-(3, 3, 4, 4, 'iJkLmN5o6p'),
-(4, 4, 4, 1, 'oPqRsT7u8v'),
-(5, 5, 4, 3, 'uVwXyZ9a0b'),
-(6, 6, 4, 2, '1a2b3c4d5e'),
-(7, 7, 4, 4, '4d5e6f7g8h'),
-(8, 8, 4, 0, '7g8h9i0jkl'),
-(9, 9, 4, 3, 'jKlMn0pQrS'),
-(10, 10, 4, 2, 'pQrStUvWxY'),
-(1, 2, 4, 4, 'vWxYz1a2b3'),
-(2, 3, 4, 3, '2b3c4d5e6f'),
-(3, 4, 4, 2, '5e6f7g8h9i'),
-(4, 5, 4, 4, '8h9i0jKlMn'),
-(5, 6, 4, 0, 'jKmNoPqRsT'),
-(6, 7, 4, 1, 'qRsTuVwXyZ'),
-(7, 8, 4, 2, 'wXyZ1a2b3c'),
-(8, 9, 4, 3, '2b3c4d5e6f'),
-(9, 10, 4, 4, '5e6f7g8h9i'),
-(10, 1, 4, 1, '8h9i0jKlMn');
+INSERT INTO results (quiz_id, user_id, name, correct_answers, url_key) VALUES
+(1, NULL, 'test_user', 3, 'aB7xY9z2w1'),
+(2, NULL, 'test_user', 2, 'cDeFgH3i4j'),
+(3, 3, 'test_user', 4, 'iJkLmN5o6p'),
+(4, 4, 'test_user', 1, 'oPqRsT7u8v'),
+(5, 5, 'test_user', 3, 'uVwXyZ9a0b'),
+(6, 6, 'test_user', 2, '1a2b3c4d5e'),
+(7, 7, 'test_user', 4, '4d5e6f7g8h'),
+(8, 8, 'test_user', 0, '7g8h9i0jkl'),
+(9, 9, 'test_user', 3, 'jKlMn0pQrS'),
+(10, 10, 'test_user', 2, 'pQrStUvWxY'),
+(1, 2, '', 4, 'vWxYz1a2b3'),
+(2, 3, '', 3, '2b3c4d5e7f'),
+(3, 4, '', 2, '5e9f7g8h9i'),
+(4, 5, '', 4, '8h9l0jKlMn'),
+(5, 6, '', 0, 'jKmNoPqRsT'),
+(6, 7, '', 1, 'qRsTuVwXyZ'),
+(7, 8, '', 2, 'wXyZ1a2b3c'),
+(8, 9, '', 3, '2b3c4d5e6f'),
+(9, 10, '', 4, '5e6f7g8h9i'),
+(10, NULL, NULL, 1, '8h9i0jKlMn');
