+++ views/results.ejs
@@ -19,9 +19,9 @@
 
   <body>
     <h1>Results! 🥔</h1>
-    <h1><label id="title"></label></h1>
-    <label id="name"></label>
-    <label id="correct-answers"></label>/
-    <label id="total-questions"></label>
+    <h1><label id="title"><%= results.title %></label></h1>
+    <label id="name"><%= results.name %></label>
+    <label id="correct-answers"><%= results.correct_answers %></label>/
+    <label id="total-questions"><%= results.total_questions %></label>
   </body>
 </html>
